Правки по заданию
